+++ examples/js/example.js
@@ -1,9 +1,9 @@
-const { lighten, darken, colorful, rotate, mix } = require('../../dist');
+const { lighten, darken, desaturate, saturate, rotate, mix } = require('../../dist');
 
 const baseColor = '#2b99ea'; console.log(baseColor);
 const darkenColor = darken(baseColor, 15); console.log(darkenColor.hex);
 const lightenColor = lighten(baseColor, 15); console.log(lightenColor.hex);
-const lessColorfulColor = colorful(baseColor, -15); console.log(lessColorfulColor.hex);
-const moreColorfulColor = colorful(baseColor, 15); console.log(moreColorfulColor.hex);
+const lessSaturatedColor = desaturate(baseColor, 15); console.log(lessSaturatedColor.hex);
+const moreSaturatedColor = saturate(baseColor, 15); console.log(moreSaturatedColor.hex);
 const rotatedColor = rotate(baseColor, -90); console.log(rotatedColor.hex);
 const mixedColor = mix(baseColor, '#f00', 0.5); console.log(mixedColor.hex);