+++ rollup.config.js
@@ -1,22 +1,20 @@
-import path from "path";
-import glob from "glob";
-import typescript from "rollup-plugin-typescript2";
-import { terser } from "rollup-plugin-terser";
+import path from 'path';
+import glob from 'glob';
+import typescript from 'rollup-plugin-typescript2';
+import { terser } from 'rollup-plugin-terser';
 
-const getRollupPluginsConfig = (compilerOptions) => {
-  return [
+const getRollupPluginsConfig = (compilerOptions) => [
     typescript({
-      tsconfigOverride: { compilerOptions },
+        tsconfigOverride: { compilerOptions },
     }),
     terser({
-      ecma: 5,
-      module: true,
-      toplevel: true,
-      compress: { pure_getters: true },
-      format: { wrap_func_args: false },
+        ecma: 5,
+        module: true,
+        toplevel: true,
+        compress: { pure_getters: true },
+        format: { wrap_func_args: false },
     }),
-  ];
-};
+];
 
 // Find available plugins
 const files = [
@@ -25,36 +23,33 @@ const files = [
         name: 'index',
         exports: undefined,
     },
-    ...glob.sync("./src/plugins/*.ts").map((input) => ({
+    ...glob.sync('./src/plugins/*.ts').map((input) => ({
         input,
         name: `plugins/${path.parse(input).name}`,
-        exports: "default",
+        exports: 'default',
     })),
 ];
 
-const outputs = files.reduce((array, { input, name, exports }, idx) => {
-    return [
-        ...array,
-        {
-            input,
-            output: {
-                file: `dist/${name}.mjs`,
-                format: "es",
-            },
-            plugins: getRollupPluginsConfig({ declaration: name === 'index' }),
+const outputs = files.reduce((array, { input, name, exports }, idx) => [
+    ...array,
+    {
+        input,
+        output: {
+            file: `dist/${name}.mjs`,
+            format: 'es',
         },
-        {
-            input,
-            output: {
-                file: `dist/${name}.js`,
-                format: "cjs",
-                exports,
-            },
-            plugins: getRollupPluginsConfig({ declaration: false }),
-        }
-    ];
-}, []);
-
+        plugins: getRollupPluginsConfig({ declaration: name === 'index' }),
+    },
+    {
+        input,
+        output: {
+            file: `dist/${name}.js`,
+            format: 'cjs',
+            exports,
+        },
+        plugins: getRollupPluginsConfig({ declaration: false }),
+    },
+], []);
 
 // Bundle both formats according to NodeJS guide
 // https://nodejs.org/api/packages.html#packages_approach_2_isolate_state