+++ src/index.ts
@@ -1,2 +1,1 @@
 export * from './lib/operations';
-export * from './plugins/less';